Trigger the `form_answers.submitted` change on submission.
The `form_answers.submitted` attribute is used in couple places to differ the moment when the User submits the form. Actually also it should be triggered by Admin/Assessor change with states select box. Some logic is coupled with the form_answers attribute so probably it's reasonable to keep it for now and think about possible refactorization.