Make review functionality more closed.
Found that Assessor is able to review the all User forms even if is not
assigned to that case and is not a lead either. Added some restrictions to
reviewing the resources also as It was designed initially to use for the Admin,
but in case of Assessors we should be more suspicious.