[Business Forms] Trade, Innovation and Development
(trello task)[https://trello.com/c/ZbzL7sni/114-additional-validation-for-business-forms]

If the applicant chooses a date on or between 1 January and 30 September
inclusive in answer to C2 (year end date) then the years displayed in subsequent questions
should be 2015 for the final year and then going backwards consecutively.

But if the applicant chooses a date in Question C2 on or between 1 October and 31 December
inclusive in answer to this question then the years should be 2014
for the final year and then going backwards consecutively.