[Non Shortlisted Mailer] type fixes
[Trello Story](https://trello.com/c/taVpyygH/230-qae-changes-to-the-not-shortlisted-email)

In the first paragraph you have missed out the year i.e. 2016. Please see my orginal attachment for the position.
In the fourth paragraph please delete the comma after the word "note"
The fifth paragraph does not need to be in bold and also the comma at the end of the sentence should be a full stop