[Jquery File Upload] issues with 'incorrect' handling of failed callback
[Trello Story](https://trello.com/c/AWBB8b9q/459-investigate-upload-issue-with-audit-certificate-for-qa0494-16i)

Once we pass:
```
forceIframeTransport: true
```
which we need for IE support -> then Jquery File Upload handles error responses incorrectly

I debugged it and found that for FileUpload
```
422 Unprocessable Entity
```
then it comes with status = 200, which is wrong.

In other words: plugin things in this case that's all good, so that we see 'undefined' message
                instead of proper error messages.

It's plugin bug.
After researching and bunch of local tests I found that it's possible
to do handling via 'always' callback, where I can check response and then
call 'upload_done' or 'failed' methods manually.