[Winners Resources] added Queen's Awards for Enterprise Press Book 2016
Could I also ask you to make a change to the winner's resources page?
Please see my attachment. Underneath "Download The Queen's Awards for Enterprise Winners' Manual 2016"
and that attachment include a new heading:

"Download The Queen's Awards for Enterprise Press Book 2016"

Then add a note which says "Please do not publicise the information in the press book
until 21 April as it is under embargo."

[Trello Story](https://trello.com/c/KkDCv7N3/322-qae2015-update-winners-resources-page)