Swap strategy for database_cleaner
The documentation says:

For the SQL libraries the fastest option will be to use :transaction as
transactions are simply rolled back. If you can use this strategy you
should.

only for js tests we change to truncation since capybara will run in
another thread and will not share the connection to rollback a
transaction.