Updating designs to match new designs by GDS
- Updated alpha/beta colors
- Updated checkboxes and radio buttons throughout USER flows.
- Due to changing Simple Form markup style, admin and assessors had to
tweak styling in order to make the checkboxes look the same as before

The change in the markup of Simple Form is due to the fact that we were
using the nested style, which puts the checkboxes and radio buttons
inside a label along with the accompanying text.

But to make the checkboxes/radio buttons as GDS designed, they need to
be before a label with the text inside. Therefore, affecting the default
config, affected admin and assessor screens, making the checkboxes look
outside the containers.