[FORM ANSWER ATTACHMENTS] corrected displaying of virus scanner check message
**WE HAVE ISSUE NOW:**

Once Admin or Assessor uploads attachment to application via 'Attach document' Then
We are displaying 'Not passed virus scanner check' message.
This message is bit confusing.

**WHAT THIS PR DOES?**

We are adding advanced check on views.
So that if file is in 'pending' or 'scanning' status, THEN:
we are displaying 'Scanning of '#{filename}' file' message.

[TRELLO STORY](https://trello.com/c/plV8rVuB/1888-qae17-attach-a-file-to-littlepods-application-and-investigate-why-it-is-failing-vigilion-scanning-on-upload)