[SECURITY UPGRADES] nokogiri to '1.8.1'
https://snyk.io/vuln/SNYK-RUBY-NOKOGIRI-20432?utm_source=slack

nokogiri (鋸) is an HTML, XML, SAX, and Reader parser, with the ability to search documents via XPath or CSS3 selectors.

Affected versions of the package are vulnerable to many vulnerabilities, including Arbitrary Code Execution and Denial of Service (DoS), and Sensitive Information Exposure. Nokogiri bundles the libxml2 library, which is vulnerable in versions below 2.9.5.

[TRELLO CARD](https://trello.com/c/A7Ur84aL/1929-qae17-tech-debt-new-vulnerability-in-gem-nokogiri-15)