Fixing autosave bug in non-editable form
When assessors were going through the application, whever they changed
steps, the autosave functionality was being called, and when that
happens and FAILS, the behaviour implemented is to refresh the page.

The fix was to simply check if the page was in read-only mode, and use
the same autosave disabling mechanism as when there's multiple
collaborators.