Fixing Dockerfile in order to fix deployments
In recent versions of Docker redundant (or "empty") COPY statements
cause an issue when building images.

In this case, copying the whole "." folder to /app, and then copying
both Gemfile and Gemfile.lock again to that folder caused a state of "no
change", and these layers are skipped.

The solution from Convox team was to move the two COPY statements
related to Gemfiles above the whole folder copy in order to never have
an "empty" copy again.

P.S: I do believe we could remove those two COPY statements, but to make
the least amount of changes to fix, I'm gonna leave them there, just in
case...