Fixing validation issue in A6.1 question
Issue has been in the app for a very long time.
Vasili spotted and debugged most of this issue.
I'm just tidying things up and committing.
Error containers were not present in the award type questions
And errors were not being cleared for answers added via AJAX, like the
award type of questions.