✨ Adding "Set A" Q&As to HTML and PDF forms
When applying for a PO award, an applicant must state whether they're
entering their entire organisation or a smaller initiative. Depending on
the user's selection, a different set of questions and answers must be
displayed (set "A" or set "B"). This commit introduces the questions and
answers for Set "A", when a user is submitting an initiative for
consideration.