✨ Adding "Set B" Q&As to HTML and PDF forms
When applying for a PO award, an applicant must state whether they're
entering their entire organisation or a smaller initiative. Depending on
the user's selection, a different set of questions and answers must be
displayed (set "A" or set "B"). This commit introduces the questions and
answers for Set "B", when a user is submitting an organisation for
consideration.