✨ Restrict number of checkbox selections
Questions B2a and B2b require the user to choose options from a list of
checkboxes. The user is permitted to choose a maximum of three options,
though the existing implementation of our `checkbox_seria` component
does not provide for this restriction. This commit adds client-side and
server-side validations to restrict the number of selections made by the
user.