♿️ Add PDF guidance to help user answer correct Qs
When submitting an application for a PO award, you're required to
specify whether you're submitting an initiative or an entire
organisation. In the HTML form, choosing your submission type
automatically updates the form below, making it obvious which questions
you need to complete. This commit updates our `options` helper to take
(optional) `pdf_context_for_option` parameters, which when provided, are
rendering beneath each option in the PDF form.