♻️ Move financial question to be PO specific
As part of the Promoting Opportunities (PO) eligibility survey, we
require users to confirm that they can provide 3 years of financial
information. Initially I implemented this as a new "basic" question,
meaning it'll be displayed in the eligibility surveys for *any* award,
though it needs to be scoped purely to the PO survey. This commit
moves the question to be part of the PO survey, exclusively.