✨ Add helper for comment text in HTML and PDF form
This commit introduces a new `comment` helper that can be used in your
form definitions to add textual context _between_ form questions.
Previously, `content` and `pdf_context` could only be supplied as part
of a question, rather than as a stand-alone element on the page.

This tool allows us to add guidance copy in the PDF form, helping
signpost applicants to the question they should be answering based on
whether they're submitting their initiative(s) or organisation.