🐛 Fix gap that could appear between form questions
As part of testing, @nirali.patel spotted an unsightly gap between two
questions in the QAE PO form. After some digging, it transpires that
client-side validations were adding the `field-with-errors` class to
hidden questions, thus adding multiple 2.45em margins to HTML divs which
combined to give the unsightly gap.

This change simply updates our CSS selector to only render the 2.45em
margin if the fieldset is a child of a visible question (having the
`show-question` class).