🐛 Fix missing award name in validation error
As part of testing, @nirali.patel spotted that selecting an ineligible
answer in the eligibility survey renders as an incomplete errors on
screen:

```
This excludes you from entering the
```

On closer inspection, each of the award eligibility classes is supposed
to override the `award_name` method, though this has been overlooked
and, as such, the award name simply returns an empty string, as defined
in the parent `eligibility` class.

This commit adds the missing `award_name` method to each of the award
eligibility classes.