🐛 Fix rendering of errors for category selection
One of the PO eligibility questions requires the user to choose whether
they wish to submit an organisation or initiative for award
consideration. This field is compulsory as enforced by our server-side
validation. Unfortunately, the corresponding `.slim` template wasn't
checking for errors when rendering the question, hence we'd simply
reload the page and hope the user connected the dots. This commit
updates our custom view to display errors to the user.

Content: We can't use simple_form for this particular partial due the
special requirement of needing some context copy to be rendered
*between* each radio button option.