🔥 Remove validation that's no longer required
In previous versions of the PO eligibility survey, we've validated that
the organisation type must be supplied and must not be a charity. We
enforced this through a custom `not_nil_or_charity` validator.

This year, "charity" does not feature as an option, thus we can use a
simple `not_nil` validator and remove the `not_nil_or_charity`
validator entirely.