🚸 Add header hints to HTML as well as PDF forms
UX feedback has shown that users sometimes get confused by our form
headings, assuming the header's posing a question that needs to be
answered. We've historically clarified expectations in the *PDF* version
of the application form by rendering a header hint beneath each heading,
though this hint never made its way into the HTML version.

This commit updates the HTML template to show the header hint and also
re-arranges the render order in the PDF such that the hint is always
directly beneath the header title.