🚸 Render conditional hints before sub_titles
We recently refactored a couple of questions in Section B of the
Promoting Opportunities (PO) form to make use of the `sub_title` method
on the `TextareaQuestion` class. This worked as expected in the HTML
form, though in the PDF form, the "conditional hint" ("Answer this
question if...") was rendering *after* the `sub_title` copy, which is
inconsistent with other questions and confusing to read.

After chatting with Nirali, we've decided that the conditional hint
should always be rendered after the primary question text, before any
`sub_title` or `context` copy.