✨ Automatically set region based on county
Feedback from QAE suggests that, in some cases, applicants aren't sure
which region to select for their organisation's address. To avoid
applicant confusion and make application data more consistent, QAE have
asked that region be automatically set based on the user's chosen
county. This commit adds a new helper module that defines the mappings
between county and region. We add a before_save to the `FormAnswer`
class so that the region is automatically set/update based on the user's
latest county.

Finally, we update the AddressQuestion class to read the list of
counties from the helper module, meaning we need only maintain a single
list of counties in the UK.