♻️ Rename principal_address_4 to principal_county
Within FormDocumentPicker there are several methods to return the
applicant's address information, reading from different database fields
depending on whether the form is for a business or non-business award.

Confusingly, the method to retrieve the applicant's county is named as
`principal_address_4` rather than `principal_county`, making it unclear
exactly the method is going to return to you. Let's boy-scout and
re-factor this method to have a clearer name 🏕.