🚸 Move Region column in All Entries CSV
I have been asked to add an applicant's region to the "All Entries" CSV,
though after looking in the code, I've found that we _already_ export
this information to the CSV, though it's hard to spot as it's not
rendered alongside the other address information.

This commit moves the column to render directly after the applicant's
county, as requested in the Trello card.

https://trello.com/c/rIESEyJn/1069-qae0620-all-forms-as-qao-i-want-to-see-distinct-columns-for-county-and-region-in-the-entries-report-so-that-i-can-determine-whic