✅ Fix failing specs by updating RegionHelper
Previously, the list of regions available for selection by the user was
defined in a constant on the `AddressQuestionDecorator` class. We have
now offloaded the maintenance of this list to a `RegionHelper` class,
though one our `.slim` partials was still referencing the (now deleted)
constant. This commit fixes the failing specs by updating the
problematic view partial.