✅ Update specs to exercise renamed method
As part of an earlier commit, I refactorted the (confusingly named)
method `principal_address4` to instead be called `principal_county`,
since this is the value the method actually returns. Unfortunately, due
to what must have been a git mishap, the associated spec changes didn't
get committed. This commit updates the specs to exercise the renamed
method.