🗂 Add new LLS data-sharing boolean to Users table
This commit adds a new boolean field to our Users table which dictates
whether the user is happy for their personal details to be shared with
Lord-Lieutenants for the purpose of application processing.

The default value for this field is deliberately `nil`, allowing us to
detect whether the user has submitted their preference (`true` or
`false`). Users with a `nil` value will be prompted to choose their
preference during the on-boarding flow (for now users) or from a new
"Additional contact preferences" page (for existing users logging in).