✨ Add new "Additional Contact Preferences" page
This commit adds a new "Additional Contact Preferences" page to which
existing users will be redirected after login. This page confirms that
the returning user is happy for their personal details to be shared with
Lord-Lieutenants during application review.

Users are required to provide a value to this question before they can
proceed, enforced by a new `before_action` on `ContentOnlyController`.

Wondering why the validation is in the controller, rather than the
model? Well, we don't want to enforce that
`agree_sharing_of_details_with_lieutenancies` is always either `true` or
`false,` since the initial value is (correctly) set to `nil` until the
user makes a choice. We don't want new User records to be invalid until
the user completes the setup journey.

Next I tried to implement and inclusion validation that only fires when
`agree_sharing_of_details_with_lieutenancies` is changed, but that means
submitting an empty form will change
`agree_sharing_of_details_with_lieutenancies` from `nil` to `nil`, hence
the validation does not trigger.