🗂 Contact preferences default to false, not true
This commit introduces a database migration to change the default values
of several DB columns from `false` to `true`.

GPDR legislation mandates that contact preferences must default to
false, giving the user to opt-in if they desire. It's important we make
this change before new applications open up for the year.

To ensure our test-suite continues to pass, we introduce a new
`agreed_to_be_contacted` trait that can be used whenever we wish to
simulate contacting the user.