🐛 Always display account owner's details
Until now we've only displayed the account owner's details as part of
the longer list of account collaborators. This means, if an account
doesn't have any collaborators, the owner's details haven't been
displayed.

This commit removes the conditional logic to ensure that an account
owner's details are always displayed, regardless of whether any other
collaborators exist.

We also special case the account owner's role in the UI, ensuring it
displays as 'Account owner', rather than 'Admin and collaborator'.

https://trello.com/c/dLYgcvRB/1097-qaeimp-collab-1-if-there-are-no-collaborators-force-user-to-add-one-when-logging-in