🐛 Skip collaborator checks for admins/assessors
As part of routine QAE support we've spotted that QAE admins are unable
to use the "Edit Application" feature for organisations that have zero
collaborator users. This makes little sense as it's not the
responsibility of the QAE admin to add/remove collaborators and they
should still be able to make amendments to any existing application(s)
in the account.

This commit skips over the collaborator presence checks if you're logged
in as a QAE admin or assessor.