💄 Add missing grey box around feedback sections
Each feedback section should be rendered within a bounding grey box,
making it easier to spot where one set of feedback ends and another
begins. This is the pattern for all of the input areas on the page.

This commit adds the missing `form-group` and `form-container` divs that
are required to render section of feedback within it's own grey box.

https://trello.com/c/MQOGvtRP