✨ Implement descriptive text for audit logs
There are several points within the QAE application where audit events
can be rendered to the user interface. To keep our code DRY and ensure
consistency across different pages, let's centralise the logic for
rendering an event as a descriptive sentence.

This commit defines a `to_s` method on the `AuditLog` model, as well as
several (private) helper methods for building up the sentence.

https://trello.com/c/YQlNYgcA/1160-qaeimp-last-updated-column-on-applications-page