♻️ Remove AJAX from submission of appraisal form
We recently received some feedback from the QAE team about
"intermittent" or "random" behaviour copying assessor comments from the
appraisal form into the "feedback" section of the page. We suspect one
source of confusion could be our use of AJAX to `POST` the appraisal
form to the server, which updates the persisted `feedback` in the
back-end, but won't re-render the feedback in the DOM. To the user, it
now seems that something is amiss as the submitting appraisal form
(seemingly) hasn't been pulled into the feedback section, as they
expect.

This commit updates the appraisal form to submit using regular HTTP,
rather than AJAX, meaning the DOM will reload in the user's browser.
Allowing the DOM to reload will trigger a re-rendering of the feedback
section, ensuring the submitted appraisal notes are now visible to the
user in the feedback section.

https://trello.com/c/MQOGvtRP/1181-qaeimp-case-summary-primary-assessor-feedback-not-consistently-pulled-through-ex-qa0606-21i