✨  Rake task to create AuditLogs from PaperTrail
Until now we've been using our `AuditService` to dynamically create
(in-memory) `AuditLog` records for all `PaperTrail::Version` records.
Whilst this approach has worked well, it's proving difficult to
optimise at a database level, giving rise to N+1 queries.

This commit adds a one-off Rake task that can be run as a one-off to
create (persisted) `AuditLog` records for all `PaperTrail::Version`
records in the database. This means we'll no longer need to synthesise
the records on the fly and we use eager_loading as the records are read
from the database.

https://trello.com/c/YQlNYgcA/1160-qaeimp-last-updated-column-on-applications-page