🔥 Remove `AuditService` and all references to it
In a previous commit we added a Rake task that creates `AuditLog`
records for all `PaperTrail::Version` records in the database. Now that
we have these records available in the database, we no longer need to
synthesise them on-the-fly, which was the main role of our
`AuditService` object.

This commit removes our (now redundant) `AuditService` object and
updates the calling code to read the `audit_log` records directly from
the database.

https://trello.com/c/YQlNYgcA/1160-qaeimp-last-updated-column-on-applications-page