✨ Reinstate "Last updated" column
The main motivation for improving the audit functionality in QAE is to
make it easier for admins/assessors to see which applications have
recently been updated. Integral to this feature is the ability to sort
the list of applications by their last update timestamp.

This commit reinstates the "Last Updated" column and introduces some
extra decorator methods for getting the most recent update time and
updater name.

https://trello.com/c/YQlNYgcA/1160-qaeimp-last-updated-column-on-applications-page