♻️ Add new `AuditEvent` object for audit history
As part of a recent code review, @dreamfall highlighted that, rather
than creating (unpersisted) `AuditLog` objects from
`PaperTrail::Version` objects, we should probably create a new
overarching object that represents either an `AuditLog` record or a
`PaperTrail::Version` record.

This commit adds a new `AuditLog` object that's built a runtime from
`PaperTrail::Version` records and `AuditLog` records. This offers a
clean separation of concern between the events we're displaying and the
database tables in which they're saved.

https://trello.com/c/YQlNYgcA/1160-qaeimp-last-updated-column-on-applications-page