🐛 Handle logging without authenticated user
We're currently seeing a failing spec due to our attempts to log updates
to the `press_summary` resource. Until now, we've assumed that changes
to a resource require a user to be logged in, although the
`press_summary` resource doesn't actually require authentication since
it'll often be completed by someone who doesn't have a QAE login. The
URL does, however, contain a secret token, so the public URLs aren't
easily guessable.

This commit refactors the `current_user` helper method so that it's no
longer defined in multiple places. If an authenticated user can't be
found, we'll log the action against a "dummy user", instead.

https://trello.com/c/YQlNYgcA/1160-qaeimp-last-updated-column-on-applications-page