♻️ Change collaborators check from 3 to 6 months
In a recent(ish) update we added some new logic to ensure that users
review (and potentially amend) their account collaborators every 3
months. Unfortunately, this change in behaviour generated a spike in
support requests for the QAE team as users were confused by the forced
redirect and didn't understand/notice the instructions to review and
confirm their collaborators on screen.

This commit bumps the collaborators grace period from 3 to 6 months,
thus reducing the frequency with which QAE users will be prompted to
review their collaborator users. Furthermore, extending this period
should reduce the number of users seeing the prompt during busy periods
(for example, around the submission deadline).

https://trello.com/c/BtUgUjoY/1186-prompt-users-to-review-their-collaborators-every-6-months-rather-than-3-months