🐛 Audit certificate must be downloaded as PDF
During internal testing we spotted that merely visiting the
`/audit_certificates` page was enough to log that the audit certificate
had been downloaded. In reality, the user hasn't downloaded the
certificate until they've followed the link to download the document as
a PDF.

This commits add a conditional check to our `before_action` so we'll
only call `log_event` when the `show` action is requested with a request
format of `pdf`.