🐛 Only log palace invite when update succeeds
During internal testing we spotted that attempts to update palace
attendee details will be logged, even when the requests failed due to
validation errors.

This commit moves our call to `log_event` into the controller action so
it'll only be called if/when a call to `@invite_form.update` succeeds.