🐛 Only log press summary when update succeeds
During internal testing we spotted that attempts to update a press
summary will be logged, even when the requests failed due to validation
errors.

This commit moves our call to `log_event` into the controller action so
it'll only be called if/when a call to `@@press_summary.update`
succeeds.