📝 Update non_shortlisted email copy
Each year QAE send out emails to notify applicants who, unfortunately,
haven't been short-listed for an award. Understandably, the content of
this email evolves over time, presumably to better set applicant
expectations and avoid a surge of support requests.

This commit updates the content of the mailer, as well as the on-screen
preview, to match the latest copy provided by the client.

https://trello.com/c/I7q6ywU1/1185-qae1020-update-the-content-of-the-email-sent-to-not-shortlisted-applicants