🐛 Separate data export logs from data update logs
We've recently overhauled the data auditing functionality of QAE to
capture many more admin/assessor/applicant events. Historically, the
only captured events related to the exporting of applicant data, and
this is something we continue to capture for data protection purposes.

When we introduced new`AuditLog` records, we added a helpful
`data_export` scope to help separate the export logs from the
application update logs. Using this helpful scope, we could ensure that
application updates didn't appear in the "Data Export Log" page.
Unfortunately, we didn't create the inverse scope which means, until
now, exporting an application as a PDF file has been treated as a change
to an application, which is incorrect.

This commit introduces a new `data_update` scope that can be used to
retrieve only the `AuditLogs` that represent a change (update) to an
application.

https://trello.com/c/YQlNYgcA/1160-qaeimp-add-a-last-updated-column-on-applications-page-and-add-logging-actions-throughout-the-app