📝 Update shortlisted email copy
Each year QAE send out emails to notify applicants who've successfully
been shortlisted for award, requesting that they login to their QAE
account, download, complete and re-upload a "Verification of Figures"
AKA "External Accountants Report" document.

This commit updates the content of the mailer, as well as the on-screen
preview, to match the latest copy provided by the client.

https://trello.com/c/MIkaqIR3/1190-qae1020-update-the-content-of-the-email-sent-to-shortlisted-applicants