✅ Handle situation where deadline is `nil`
In a recent update we made changes to the guidance that's displayed to
users when they must download, complete and re-upload an
`audit_certificate`. Part of the copy displays the submission deadline,
calling `strftime` to display the deadline in a human-readable format.
Unfortunately, it's possible that a deadline hasn't yet been set by
admin, causing a runtime error when we try to run `strftime` on a `nil`
deadline.

This commit makes use of existing helper method on `Settings` which
gracefully handles undefined deadline values.