✨ Add new ListOfProcedures model
For this new award year, QAE are adding an extra verification step for
applicants who have been short-listed for an award. Going forward, in
addition to uploading the regular "Verification of Figures" document,
now known as an "External Accountant's Report", applicants must also
upload a "List of Procedures" document.

This commit introduces a new `ListOfProcedure` model as well as the
database migrations needed for persistence.

https://trello.com/c/UGpnY8ei/1194-qaeimp-add-an-upload-list-of-procedures-feature