🔥 Temporarily remove VOCF status
Historically, verification of commercial figures (VOCF) involved the
upload of a single document. Going forward, short-listed applicants must
upload a verification document and a secondary "List of procedures". The
addition of a second document raises a new question; what status should
be shown when on 1/2 documents has been uploaded?

This commit temporarily removes the status indicator for "Verification
of Commercial Figures" whilst we wait for QAE to advise on if/how this
should change since the addition of a second required document.

https://trello.com/c/UGpnY8ei/1194-qaeimp-add-an-upload-list-of-procedures-feature