💄 Fix deadline styling for longer descriptions
In a previous commit we updated our locales to provide more information
about the "Verification of Commercial Figures" process, explicitly
naming the 2 documents that need to be uploaded by the user.
Unfortunately, adding more descriptive text introduced some unsightly
line-wrapping that needs to be addressed.

This commit redefines the styling for the `data-time-text` class,
removing the hard-coded `width` property and adding some `padding`, both
of which work allow longer descriptive text to render correctly for
application deadlines.