♻️ Check list of procedures when sending reminders
Until now the "Verification of Commercial Figures" process have involved
the upload of a single document, saved internally as the
`audit_certificate` record. This year, QAE have added a new
`list_of_procedures` document that must also be uploaded as part of the
verification process.

This commit updates our reminder email logic to check for both the
`audit_certificate` *and* `list_of_procedures`, when deciding whether
the applicant should be sent a reminder email for the upload of their
documents.

https://trello.com/c/9QZaxIuy/1192-qaeimp-update-the-content-of-the-email-sent-to-remind-applicants-to-upload-their-external-accountants-report-and-the-supplementa