🐛 Attempt to fix flaky Capybara specs
Since migrating from GitLab CI to GitHub Actions, we've spotted some
flaky capybara specs that intermittently fails in our CI pipeline,
causing our deployments to fail and needing to be retried.

After reading through the spec, I've found that we're trying to interact
with HTML elements that are originally hidden and become visible as
part of a Javascript interaction. Ordinarily, Capybara is smart enough
to wait for Javascript interactions to complete and will look for the
element multiple times, though I suspect this courtesy doesn't extend
to `within` blocks, meaning the element to which we're scoping our
interaction isn't always visible and thus causing a race condition.

This commit forces Capybara to wait for the element to become visible
(within a sensible period of time) using the `expect` method, which has
the nice behaviour of silently retrying if/when a selector doesn't match
any visible elements.

https://trello.com/c/HG1f8xNL/1196-qae1020-dig-into-the-flaky-spec-that-intermittently-causes-our-ci-pipelines-to-fail