🐛 Correctly log verification document deletion
Now that users are allowed to remove their uploaded verification
document(s), it's important that an accurate audit trail is captured,
else there'll be inevitable confusion.

Whilst testing the new removal functionality I realised that removal of
the "list of procedures" document was being incorrectly logged as a
document upload, since until recently, only the `create` action existed
on the controller.

This commit updates both of our document controllers to ensure accurate
audit logs are written to the database.

https://trello.com/c/UGpnY8ei/1194-qaeimp-add-an-upload-list-of-procedures-feature