♻️ Show "remove" link immediately after doc upload
In a previous commit we implemented "remove" functionality in our
"Verification of Commercial Figures" page, allowing users to remove the
document(s) they've uploaded for review. During testing, Stephanie
highlighted that the "remove" button doesn't appear when you first
upload a file and instead is only visible after a page reload.

This commit updates our view to always render the "remove" button,
though it's hidden by default and is subsequently made visible by the
(existing) Javascript that updates the UI in response to a file upload.