🐛 Handle audit events from deleted users
This morning we saw a Sentry error (linked below) where we're an
`AuditEvent` object had a `subject` value of `nil`. Upon closer
inspection, I've found that 74 `PaperTrail::Version` records were
created by an Assessor that no longer exists in the database, hence
a `nil` value is being set for the `subject` association.

This commit ensures that a `subject` is always set by falling back to
our `dummy_user`, if/when the "real" subject can't be retrieved from the
database.

https://sentry.io/organizations/bit-zesty-client-apps/issues/2020379239/?referrer=slack

https://trello.com/c/Wa3DxWGF/1209-qae1120-sentry-error-undefined-method-fullname-for-nilnilclass