📈 Track when an assessor's note can't be saved
We've recently received a support request from QAE reporting that, for a
handful of assessors, submitted case summaries (a type of
`AssessorAssignment`) haven't saved correctly. So far we've been unable
to verify or replicate the issue.

This commit adds a Sentry notification if/when an `AssessorAssignment`
object fails to be successfully updated.

https://app.asana.com/0/1183526305150907/1199195899126653