✨ Accountant's Report: Show filename in DL link
As part of our ongoing efforts to improve the QAE UX, we'd like the
download links for user-uploaded documents to feature the name of the
document, making it easier for the user to verify what they've uploaded
and to understand what they're about to re-download.

This commit updates the download link for the External Accountant's
Report.