💄 Add download icons and remove bullet points
In a previous commit we updates the Verification of Commercial Figures
page to display download links that contain the name of the
user-uploaded files. This commit further improves the UX of the page,
adding download icons before each hyperlink and removing the unsightly
radio buttons that otherwise clash with the download glyph.