✨ Show file name in download link (AJAX)
In a previous update we changed our server-rendered HTML links to
include the name of the user-uploaded file. This works beautifully for
files that have been previously uploaded and that the server therefore
knows the full filename, but what about files that are about to be
uploaded via AJAX?

This commit updates the javascript for handling a successful file
upload, updating the `text` of the anchor link to show the name of the
user-uploaded file.