💄 Increase vertical space between upload buttons
We've been asked to increase the spacing between the 2 upload buttons on
the Verification of Commercial figures page. The default `margin-bottom`
for `span.button-add`is `10px`, though we'd like to special-case the
buttons on this particular page, doubling the `margin-bottom` value to
`20px`.

This commit adds an extra CSS rule, scoped specifically to `span`
elements that have the `button-add` class and are nested within the
`audit-certificate-upload` form.