♻️ Add `download` attribute to download links
Based on some (excellent) advice from @mauricioac, this commit adds the
`download` attribute to our links for downloading verification
documents. Adding this attribute tells the browser to download the file,
rather than open it in a new tab, and also sets the default filename to
match that of the originally uploaded file.

References:
https://developers.google.com/web/updates/2011/08/Downloading-resources-in-HTML5-a-download?hl=en