💄 Tweak style of assessor financial verification
We've received a handful of reports from assessor users claiming the
notes they've added under the the "Financial Summary" section, whilst
reviewing an application, either haven't been or saved or have been
lost. Whilst we continue to investigate these reports, let's tweak the
UI to move the form controls to the left of the page and closer to the
confirmation radio buttons, ensuring users notice them when applying
changes.

https://app.asana.com/0/1199190913173550/1199198194706069/f