💄 Fix label overlapping with radio buttons
During testing, Stephanie spotted that controls for commercial figure
verification, whilst viewing an application that doesn't have an
associated `audit_certificate` (AKA External Accountant's Report),
renders a label that overlaps with the form radio inputs.

This commit resolves the overlapping elements by moving the label to the
top of the Financial Summary panel, rather than having it rendered by a
partial *after* the radio buttons have been rendered, which was the
cause of the overlap.