🐛 Update "All Entries" CSV to support new keys
This morning, we received a support request from QAE highlighting how,
when viewing "Recommended" applications in the "All Entries" CSV export,
the "ProductService" column was blank for "Sustainable Development" and
"Promoting Opportunities" awards.

After reading through the exporter code I've found that, in recent
years, we've subtly reworded the questions from which this column is
populated, including changing the name of the `document` hash key in
which the user's answer is saved.

This commit updates our CSV exporter logic to read from the appropriate
hash key depending on the application year being exported.