⬆️ Upgrade mimemagic to release with MIT licence
Yesterday, in an event that shook the Rails community to its core
(team) 😂, the maintainer of the `mimemagic` gem, a dependency of
`activestorage` and therefore a default dependency of `rails`, yanked
all releases prior to 0.3.6. The yanking of versions relates to a change
in software licence, though it caused chaos for Ruby developers as
their CI builds and deployments started to fail when `mimemagic` could
no longer be downloaded and installed.

This commit upgrades our codebase to `mimemagic` version `0.3.9`, the
latest release in the `0.3.x` branch, which has removed the contentious
code and is once again available under an MIT licence.

https://github.com/minad/mimemagic/issues/98